reward-server/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-cors==0.8.0
aiosignal==1.3.2
airportsdata==20250224
albucore==0.0.13
albumentations==1.4.10
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
av==14.2.0
beautifulsoup4==4.13.3
bitsandbytes==0.45.3
blake3==1.0.4
blinker==1.9.0
braceexpand==0.1.7
cachetools==5.5.2
charset-normalizer==3.4.1
click==8.1.8
clip
clip-benchmark==1.6.1
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
compressed-tensors==0.9.1
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cuda-bindings==12.8.0
cuda-python==12.8.0
cycler==0.12.1
Cython==3.0.12
datasets
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.4
depyf==0.18.0
diffusers==0.33.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docstring-parser==0.16
einops==0.8.1
einops-exts==0.0.4
exceptiongroup==1.2.2
executing==2.2.0
fairscale==0.4.13
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.14.0
fire==0.7.0
# flash-attn==2.7.4.post1 #2.6.3 works
flashinfer-python==0.2.3
flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gguf==0.10.0
gitdb==4.0.12
gitpython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
hf-transfer==0.1.9
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
icecream==2.1.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imgaug==0.4.0
importlib-metadata==8.6.1
importlib-resources==6.5.2
inflect==6.0.4
intel-openmp==2023.1.0
interegular==0.3.3
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.5
jiter==0.8.2
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.8
lark==1.2.2
latex2mathml==3.77.0
lazy-loader==0.4
Levenshtein==0.26.1
litellm==1.63.12
llguidance==0.7.4
lm-format-enforcer==0.10.11
lmdb==1.6.2
lxml==5.3.1
markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
markupsafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral-common==1.5.4
ml-collections==1.0.0
model-index==0.1.11
modelscope==1.24.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open-clip-torch==2.31.0
openai
opencensus==0.11.4
opencensus-context==0.1.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
outlines==0.1.11
outlines-core==0.1.26
packaging==24.2
paddleocr==2.9.1
paddlepaddle-gpu==2.6.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.10.0
pexpect==4.9.0
Pillow==10.4.0
platformdirs==4.3.6
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyclipper==1.3.0.post6
pycocoevalcap==1.2
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytz==2023.4
pyzmq==26.3.0
qwen-vl-utils==0.0.10
rapidfuzz==3.12.1
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.28.2
rich==13.4.2
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.4
setuptools==60.2.0
sgl-kernel==0.0.5
sglang==0.4.4.post1
shapely==2.0.7
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack_data==0.6.3
starlette==0.46.1
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.13.3
tomli==2.2.1
torchao==0.9.0
torchmetrics==1.7.1
tqdm
traitlets==5.14.3
transformers==4.28.0
triton==2.1.0
typeguard==4.4.2
typing_extensions==4.12.2
tyro==0.9.16
tzdata==2025.1
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.3
wandb==0.18.7
watchfiles==1.0.4
wcwidth==0.2.13
webdataset==0.2.111
websockets==11.0.3
werkzeug==3.1.3
wrapt==1.17.2
xformers
xgrammar
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0